+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### [Latest]
 
+- Add `--update-reference-plots` pytest flag to regenerate image-comparison reference plots
 - Fix Issues in with Matplotlib 3.11 [#372](https://github.com/umami-hep/puma/pull/372)
 - Update Docker development image [#370](https://github.com/umami-hep/puma/pull/370)